+++ app/package.json
@@ -13,7 +13,6 @@
     "@mui/icons-material": "^5.15.18",
     "@mui/material": "^5.15.18",
     "@reduxjs/toolkit": "^2.2.5",
-    "react-toastify": "^10.0.5",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
@@ -35,11 +34,14 @@
     "mobx-react": "^9.1.1",
     "prettier": "^3.2.5",
     "react": "^18.3.1",
+    "react-bootstrap": "^2.10.2",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.51.4",
     "react-redux": "^9.1.2",
     "react-router-dom": "^6.23.1",
     "react-scripts": "5.0.1",
+    "react-toastify": "^10.0.5",
+    "reactjs-popup": "^2.0.6",
     "recoil": "^0.7.7",
     "redux": "^5.0.1",
     "redux-thunk": "^3.1.0",
