+++ app/src/components/modal/ModalRemoveCartItem.tsx
@@ -1,43 +1,2 @@
-import React from 'react';
-// import "./css/styleModalRemoveCartItem.css";
-
-interface ModalRemoveCartItemProps {
-    setOpenModal: (isOpen: boolean) => void;
+export default function ModalRemoveCartItem() {
 }
-
-const ModalRemoveCartItem: React.FC<ModalRemoveCartItemProps> = ({ setOpenModal }) => {
-    return (
-        <div className="modalBackground">
-            <div className="modalContainer">
-                <div className="titleCloseBtn">
-                    <button
-                        onClick={() => {
-                            setOpenModal(false);
-                        }}
-                    >
-                        X
-                    </button>
-                </div>
-                <div className="title">
-                    <h1>Are You Sure You Want to Continue?</h1>
-                </div>
-                <div className="body">
-                    <p>The next page looks amazing. Hope you want to go there!</p>
-                </div>
-                <div className="footer">
-                    <button
-                        onClick={() => {
-                            setOpenModal(false);
-                        }}
-                        id="cancelBtn"
-                    >
-                        Cancel
-                    </button>
-                    <button>Continue</button>
-                </div>
-            </div>
-        </div>
-    );
-}
-
-export default ModalRemoveCartItem;
