+++ app/src/layout/MainLayout.tsx
@@ -9,10 +9,9 @@ const MainLayout = () => {
     return (
         <>
             <Topbar />
-            <Navbar />
+            {/*<Navbar />*/}
             <Outlet />
             <Footer />
-
         </>
     )
 }
